requirements.txt
lightning>=2.4
hydra-core>=1.3
omegaconf>=2.3
torchvision>=0.19
tensorboard
